+++ policy_sentry/shared/database.py
@@ -276,7 +276,7 @@ def build_action_table(db_session, service, access_level_overrides_file):
                     if access_level_overrides_cfg:
                         override_result = determine_access_level_override(
                             service,
-                            str.lower(action_name),
+                            action_name,
                             table["data"][i][2],
                             access_level_overrides_cfg,
                         )
@@ -320,7 +320,7 @@ def build_action_table(db_session, service, access_level_overrides_file):
                     db_session.add(
                         ActionTable(
                             service=service,
-                            name=str.lower(action_name),
+                            name=action_name,
                             description=table["data"][i][1],
                             access_level=access_level,
                             resource_type_name=resource_type_name,