+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Other Pending:
 * Terraform module updates are currently pending
 * Update the GIF and the Medium blog
 
+## Next release (Unreleased)
+* Fix for #124. Working on making sure the output is with UpperCamelCase rather than all lowercase, for human readability.
+* If `--minimize` is specified for `write-policy`, it will give me lowercase. Otherwise, it will be UpperCamelCase.
+
 ## 0.7.0.2 (2020-02-10)
 * Ignore empty entries in the yaml template